+++ tests/test_build.py
@@ -5,8 +5,15 @@ import shutil
 
 from click.testing import CliRunner
 from htmd.cli import build
+import yaml
 
-from utils import remove_fields_from_post, set_example_to_draft, SUCCESS_REGEX
+from utils import (
+    remove_fields_from_post,
+    set_example_password_value,
+    set_example_to_draft,
+    set_example_to_draft_build,
+    SUCCESS_REGEX,
+)
 
 
 def test_build(run_start: CliRunner) -> None:
@@ -313,3 +320,136 @@ def test_build_contains_favicon(run_start: CliRunner) -> None:
     with build_favicon.open('r') as favicon_file:
         contents = favicon_file.read()
     assert contents == original_contents
+
+
+def test_build_password_protect(run_start: CliRunner) -> None:
+    set_example_password_value('')
+    with (Path('posts') / 'example.md').open('r') as md_file:
+        md_str = md_file.read()
+    data = yaml.safe_load(md_str[:md_str.find('...')])
+    password = data['password']
+    assert password is None
+    result = run_start.invoke(build)
+    assert result.exit_code == 0
+    assert re.search(SUCCESS_REGEX, result.output)
+    post_path = Path('build') / '2014' / '10' / '30' / 'example' / 'index.html'
+    with post_path.open('r') as post_file:
+        contents = post_file.read()
+    with (Path('posts') / 'example.md').open('r') as md_file:
+        md_str = md_file.read()
+
+    # extract password
+    data = yaml.safe_load(md_str[:md_str.find('...')])
+    password = data['password']
+
+    assert 'Posted by <a href="/author/Taylor/">Taylor</a> on 2014-10-30' in contents
+    title = 'Example Post'
+    assert title in md_str
+    assert title not in contents
+    body = 'This is the post'
+    assert body in md_str
+    assert body not in contents
+
+    # build again and verify that the password has not changed
+    result = run_start.invoke(build)
+    assert result.exit_code == 0
+    assert re.search(SUCCESS_REGEX, result.output)
+    with (Path('posts') / 'example.md').open('r') as md_file:
+        md_str = md_file.read()
+    data = yaml.safe_load(md_str[:md_str.find('...')])
+    after_password = data['password']
+    assert after_password == password
+
+
+def test_build_password_false(run_start: CliRunner) -> None:
+    set_example_password_value('false')
+    result = run_start.invoke(build)
+    assert result.exit_code == 0
+    assert re.search(SUCCESS_REGEX, result.output)
+
+    post_path = Path('build') / '2014' / '10' / '30' / 'example' / 'index.html'
+    with post_path.open('r') as post_file:
+        contents = post_file.read()
+
+    assert 'Posted by <a href="/author/Taylor/">Taylor</a> on 2014-10-30' in contents
+    title = 'Example Post'
+    assert title in contents
+    body = 'This is the post'
+    assert body in contents
+
+    with (Path('posts') / 'example.md').open('r') as md_file:
+        md_str = md_file.read()
+    data = yaml.safe_load(md_str[:md_str.find('...')])
+    assert data['password'] is False
+
+
+def test_build_draft_password_protect(run_start: CliRunner) -> None:
+    set_example_to_draft_build()
+    set_example_password_value('')
+    with (Path('posts') / 'example.md').open('r') as md_file:
+        md_str = md_file.read()
+    data = yaml.safe_load(md_str[:md_str.find('...')])
+    password = data['password']
+    assert password is None
+    result = run_start.invoke(build)
+    assert result.exit_code == 0
+    assert re.search(SUCCESS_REGEX, result.output)
+
+    with (Path('posts') / 'example.md').open('r') as md_file:
+        md_str = md_file.read()
+    data = yaml.safe_load(md_str[:md_str.find('...')])
+    build_uuid = data['draft'].replace('build|', '')
+
+    post_path = Path('build') / 'draft' / build_uuid / 'index.html'
+    with post_path.open('r') as post_file:
+        contents = post_file.read()
+
+    # extract password
+    data = yaml.safe_load(md_str[:md_str.find('...')])
+    password = data['password']
+
+    assert 'Posted by <a href="/author/Taylor/">Taylor</a> on 2014-10-30' in contents
+    title = 'Example Post'
+    assert title in md_str
+    assert title not in contents
+    body = 'This is the post'
+    assert body in md_str
+    assert body not in contents
+
+    # build again and verify that the password has not changed
+    result = run_start.invoke(build)
+    assert result.exit_code == 0
+    assert re.search(SUCCESS_REGEX, result.output)
+    with (Path('posts') / 'example.md').open('r') as md_file:
+        md_str = md_file.read()
+    data = yaml.safe_load(md_str[:md_str.find('...')])
+    after_password = data['password']
+    assert after_password == password
+
+
+def test_build_draft_password_false(run_start: CliRunner) -> None:
+    set_example_to_draft_build()
+    set_example_password_value('false')
+    result = run_start.invoke(build)
+    assert result.exit_code == 0
+    assert re.search(SUCCESS_REGEX, result.output)
+
+    with (Path('posts') / 'example.md').open('r') as md_file:
+        md_str = md_file.read()
+    data = yaml.safe_load(md_str[:md_str.find('...')])
+    build_uuid = data['draft'].replace('build|', '')
+
+    post_path = Path('build') / 'draft' / build_uuid / 'index.html'
+    with post_path.open('r') as post_file:
+        contents = post_file.read()
+
+    assert 'Posted by <a href="/author/Taylor/">Taylor</a> on 2014-10-30' in contents
+    title = 'Example Post'
+    assert title in contents
+    body = 'This is the post'
+    assert body in contents
+
+    with (Path('posts') / 'example.md').open('r') as md_file:
+        md_str = md_file.read()
+    data = yaml.safe_load(md_str[:md_str.find('...')])
+    assert data['password'] is False